requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
flask_socketio
eventlet